capstone_project/backend/requirements.txt
# Core Framework
# Pin starlette: FastAPI 0.109.x requires starlette 0.35.x (1.x breaks APIRouter / on_startup).
starlette>=0.35.0,<0.36.0
fastapi==0.109.0
# mcp>=1.26.0 requires uvicorn>=0.31.1 (stdio server helpers).
uvicorn[standard]>=0.31.1,<0.43.0
python-dotenv==1.0.0

# LangChain & AI (relaxed versions for compatibility)
langchain>=0.2.0
langchain-openai>=0.1.0
langchain-qdrant>=0.1.0
langchain-community>=0.2.0
langchain-text-splitters>=0.2.0

# Vector Store & Embeddings
qdrant-client>=1.10.1,<2.0.0

# Database
sqlalchemy==2.0.25

# MCP stdio client (required for default real Agentic MCP; tests set USE_SIMULATED_MCP=1)
mcp>=1.26.0,<2.0.0

# Utilities
pydantic>=2.7.4,<3.0.0
pydantic-settings>=2.4.0,<3.0.0
# mcp>=1.2 requires httpx>=0.27.1; keep below 0.29 for predictable API.
httpx>=0.27.1,<0.29.0
# mcp>=1.26.0 requires python-multipart>=0.0.9
python-multipart>=0.0.9,<0.1.0

# Testing (pytest-asyncio 0.23.x breaks pytest 8 collection on Package; need >=0.24 + pytest>=8.2)
pytest>=8.2.0,<9.0.0
pytest-asyncio>=0.24.0,<0.26.0
pytest-cov==4.1.0
pytest-mock==3.12.0
faker==22.0.0

# AI Quality Testing (optional - install separately if needed)
# ragas==0.1.1
# datasets==2.16.1

# Development
black==24.1.1
ruff==0.1.14
mypy==1.8.0
